Create generic selection-header from list-header
